net: wireguard: Feed the packet directly to control interface
Things have changed in the upstream network stack and we cannot
feed the packet to network stack using net_recv_data() as that
will cause infinite loop because UDP packet handler will match
all the interfaces and we would receive the same packet in
the UDP handler for the control interface.

Solution to this is to directly feed the packet to the control
interface which will handle it directly.

Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>